requirements.txt
Flask==1.1.2
pycryptodome==3.9.7
requests==2.23.0
